// FRONTEND/chatbot-admin-panel/src/services/api/schemas/lLMModelConfigCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { LLMProviderType } from './lLMProviderType';
import type { LLMModelType } from './lLMModelType';
import type { LLMModelConfigCreateBaseUrl } from './lLMModelConfigCreateBaseUrl';
import type { LLMModelConfigCreateDefaultTemperature } from './lLMModelConfigCreateDefaultTemperature';
import type { LLMModelConfigCreateDefaultMaxTokens } from './lLMModelConfigCreateDefaultMaxTokens';
import type { LLMModelConfigCreateConfigJson } from './lLMModelConfigCreateConfigJson';
import type { LLMModelConfigCreateApiKeyPlain } from './lLMModelConfigCreateApiKeyPlain';

export interface LLMModelConfigCreate {
  /**
   * @minLength 1
   * @maxLength 255
   */
  model_identifier: string;
  /**
   * @minLength 3
   * @maxLength 150
   */
  display_name: string;
  provider: LLMProviderType;
  model_type?: LLMModelType;
  is_active?: boolean;
  base_url?: LLMModelConfigCreateBaseUrl;
  default_temperature?: LLMModelConfigCreateDefaultTemperature;
  default_max_tokens?: LLMModelConfigCreateDefaultMaxTokens;
  supports_system_prompt?: boolean;
  config_json?: LLMModelConfigCreateConfigJson;
  /** API Key en texto plano. Se encriptará antes de guardar. */
  api_key_plain?: LLMModelConfigCreateApiKeyPlain;
}