// FRONTEND/chatbot-admin-panel/src/services/api/schemas/lLMModelConfigUpdate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { LLMModelConfigUpdateModelIdentifier } from './lLMModelConfigUpdateModelIdentifier';
import type { LLMModelConfigUpdateDisplayName } from './lLMModelConfigUpdateDisplayName';
import type { LLMModelConfigUpdateProvider } from './lLMModelConfigUpdateProvider';
import type { LLMModelConfigUpdateModelType } from './lLMModelConfigUpdateModelType';
import type { LLMModelConfigUpdateIsActive } from './lLMModelConfigUpdateIsActive';
import type { LLMModelConfigUpdateBaseUrl } from './lLMModelConfigUpdateBaseUrl';
import type { LLMModelConfigUpdateDefaultTemperature } from './lLMModelConfigUpdateDefaultTemperature';
import type { LLMModelConfigUpdateDefaultMaxTokens } from './lLMModelConfigUpdateDefaultMaxTokens';
import type { LLMModelConfigUpdateSupportsSystemPrompt } from './lLMModelConfigUpdateSupportsSystemPrompt';
import type { LLMModelConfigUpdateConfigJson } from './lLMModelConfigUpdateConfigJson';
import type { LLMModelConfigUpdateApiKeyPlain } from './lLMModelConfigUpdateApiKeyPlain';

export interface LLMModelConfigUpdate {
  model_identifier?: LLMModelConfigUpdateModelIdentifier;
  display_name?: LLMModelConfigUpdateDisplayName;
  provider?: LLMModelConfigUpdateProvider;
  model_type?: LLMModelConfigUpdateModelType;
  is_active?: LLMModelConfigUpdateIsActive;
  base_url?: LLMModelConfigUpdateBaseUrl;
  default_temperature?: LLMModelConfigUpdateDefaultTemperature;
  default_max_tokens?: LLMModelConfigUpdateDefaultMaxTokens;
  supports_system_prompt?: LLMModelConfigUpdateSupportsSystemPrompt;
  config_json?: LLMModelConfigUpdateConfigJson;
  /** Proporcionar una nueva API Key para actualizarla. */
  api_key_plain?: LLMModelConfigUpdateApiKeyPlain;
}